+++ .github/workflows/ci.yml
@@ -33,10 +33,11 @@ jobs:
         shell: bash
         run: |
           set -ex
-          sudo xcode-select --switch /Applications/Xcode_12.3.app/Contents/Developer/
+          sudo xcode-select --switch /Applications/Xcode_12.4.app/Contents/Developer/
           # avoid building unrelated products for testing by specifying the test product explicitly
           swift build --product TokamakPackageTests
-          `xcrun --find xctest` .build/debug/TokamakPackageTests.xctest
+          `xcrun --find xctest` .build/debug/TokamakPackageTests.xctest ||
+            find /var/folders -iname SnapshotTests -exec cp -r {} . \;
 
           rm -rf Sources/TokamakGTKCHelpers/*.c
 
@@ -55,6 +56,13 @@ jobs:
 
           ./benchmark.sh
 
+      - name: Upload failed snapshots
+        uses: actions/upload-artifact@v2
+        if: ${{ failure() }}
+        with:
+          name: Failed snapshots
+          path: SnapshotTests
+
   gtk_macos_build:
     runs-on: macos-latest
 
@@ -64,7 +72,7 @@ jobs:
         shell: bash
         run: |
           set -ex
-          sudo xcode-select --switch /Applications/Xcode_12.3.app/Contents/Developer/
+          sudo xcode-select --switch /Applications/Xcode_12.4.app/Contents/Developer/
 
           brew install gtk+3
